Queue/requirements.txt
Flask==1.1.1
h5py==2.8.0
matplotlib==3.2.1
numpy==1.18.1
pandas==1.0.3
Pillow==7.0.0
progress==1.5
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
opencv-python
